fix(ui): added a generic error page for general errors
